app/requirements.txt
flask
RPi.GPIO
Flask-APScheduler